+++ lib/ack-tracker/tests/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_unit_test(CRITERION TARGET test_late_ack_record_container)
+add_unit_test(CRITERION TARGET test_consecutive_ack_record_container)
