+++ lib/ack-tracker/tests/Makefile.am
@@ -1,8 +1,8 @@
 lib_ack_tracker_tests_TESTS			=  \
-	lib/ack-tracker/tests/test_late_ack_record_container
+	lib/ack-tracker/tests/test_consecutive_ack_record_container
 
 check_PROGRAMS				+= \
 	${lib_ack_tracker_tests_TESTS}
 
-lib_ack_tracker_tests_test_late_ack_record_container_LDADD	= $(TEST_LDADD)
-lib_ack_tracker_tests_test_late_ack_record_container_CFLAGS	= $(TEST_CFLAGS)
+lib_ack_tracker_tests_test_consecutive_ack_record_container_LDADD	= $(TEST_LDADD)
+lib_ack_tracker_tests_test_consecutive_ack_record_container_CFLAGS	= $(TEST_CFLAGS)
